requirements.txt
accelerate
aiohttp
attrs
bottleneck
brotli
certifi
charset-normalizer
colorama
contourpy
cycler
datasets
dill
filelock
fonttools
frozenlist
fsspec
huggingface-hub
idna
jinja2
joblib
kiwisolver
matplotlib==3.10.0
mpmath
multidict
multiprocess
networkx
numexpr
numpy
packaging
pandas
pillow
psutil
pyarrow==18.1.0
pyparsing
python-dateutil
pytz
pyyaml
regex
requests
safetensors
scikit-learn
scipy
seaborn
sentence-transformers
six
sympy
threadpoolctl
tokenizers
torch
tornado
tqdm
transformers
typing-extensions
urllib3
xxhash
yarl
openai
google-genai
